Fix Xcode warnings (#545)
## Changes
* Fixes warnings generated when compiling with Xcode:
  * There was some implicit type conversion.
  * The macOS deployment target is updated to 11.0 (Big Sur).

+++ libSetReplace/Parallelism.hpp
@@ -44,7 +44,7 @@ bool isAvailable(const HardwareType& type);
 
 #ifdef LIBSETREPLACE_BUILD_TESTING
 namespace Testing {
-void overrideNumHardwareThreads(const HardwareType& type, const int64_t& numThreads);
+void overrideNumHardwareThreads(const HardwareType& type, const unsigned& numThreads);
 }
 #endif
 }  // namespace SetReplace::Parallelism
